agents/trader/requirements.txt
# NeuroFlow Agent Dependencies
# trader

# Core
neuroflow-sdk>=0.4.1

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0

# Configuration
pyyaml>=6.0

# HTTP Client
aiohttp>=3.8.0

# MCP (optional, for advanced usage)
# @modelcontextprotocol/sdk

# Development (optional)
# pytest>=7.0.0
# black>=23.0.0
# isort>=5.0.0
